Small changes to the cellResGUI to try to speed up the interaction between the keyboard and the GUI as there was a suprising amount of lag.

+++ Matt Seg GUI/@timelapseTraps/loadChannelIntoMemory.m
@@ -11,7 +11,7 @@ end
 if loadIn
     h = waitbar(0,'Please wait as the channel is loaded ...');
 
-    cTimelapse.temporaryImageStorage.images=cTimelapse.returnSingleTimepoint(1,channel);
+    cTimelapse.temporaryImageStorage.images=uint8(cTimelapse.returnSingleTimepoint(1,channel));
     maxTP=length(cTimelapse.timepointsProcessed);
     if maxTP>1
         cTimelapse.temporaryImageStorage.images(:,:,2:maxTP) = 0;
@@ -19,7 +19,7 @@ if loadIn
     
     for timepoint=2:maxTP;
         waitbar(timepoint/maxTP);
-        cTimelapse.temporaryImageStorage.images(:,:,timepoint)=cTimelapse.returnSingleTimepoint(timepoint,channel);
+        cTimelapse.temporaryImageStorage.images(:,:,timepoint)=uint8(cTimelapse.returnSingleTimepoint(timepoint,channel));
     end
     
     cTimelapse.temporaryImageStorage.channel=channel;
